Move base class to same testing directory
