+++ Testing/SystemTests/tests/qt/AlgorithmDialogsMDWorkspaceStartupTest.py
@@ -6,7 +6,7 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 from mantid.simpleapi import CreateMDWorkspace
 
-from testhelpers.AlgorithmDialogsStartupTestBase import AlgorithmDialogsStartupTestBase
+from AlgorithmDialogsStartupTestBase import AlgorithmDialogsStartupTestBase
 
 
 class AlgorithmDialogsMDWorkspaceStartupTest(AlgorithmDialogsStartupTestBase):
